Detect when the cache changes and react accordingly
We currently have a (somewhat cornercasey) situation where, if two
Engine instances handle the same knsrc file, they might easily end
up out of sync. This makes the cache handler watch the cache file
for changes made by others, and when a change is registered, the
cache gets reloaded, and any mismatches between the old and new
cache get emitted as changed entries, which the Engine then reacts
to.

This specifically will allow e.g. Discover's update notifier to learn
when an entry is changed through a Get Hot New Stuff dialog somewhere
else on the system (straightforward scenario, someone's got the
notifier telling them there's an update for some wallpaper. The
user then goes and installs another wallpaper through the dialog, and
while there notices that there is an update for another. They click
that update button, and currently that would leave the notifier out
of sync with reality (that is, the update was installed, but the
notifier would still think there was an update to be made)).