Manually track the life cycle of our kpackage runner internals
This equalises how the various bits held by the kpackage runner are
managed (the two kpackage structures are already handled manually),
which mostly makes me happy. It also means i don't have to worry about
whether or not that job is actually deleted correctly after the event
loop exits in the lambda connected to its result.