Fix double-delete crash in kpackagejob
Don't hang onto a pointer that's already been deleted, as QScopedPointer
will just delete that for us. The semantics of
KPackage::PackageStructures is that someone else /might/ own it, while
the semantics of KPackage::Package is that we definitely own it, so we
need to track the one we don't really know about.

BUG:427910