Reuse QWidgets dialog
Otherwise the engine would make a new network request, this is quite
wasteful and lets the user wait a moment.
Also this appears to fixes the crash, you can try this out by quickly
opening and closing the dialog.

Additionally this makes it consistent with the QtQuick one.

From the stacktrace I got this seems to fix BUG 429302.

BUG: 416328
BUG: 429302
FIXED-IN: 5.77