Create wrapper class for QML dialog
This will allow KCMs and applications to utilize the
QtQuick dialog without needing to be rewritten in QML.

Also some signals are directly forwarded from the engine,
this will make the consumers more flexible when
reacting to installed/updated/uninstalled entries or
the adoption command.

Once the existing usages have been ported we can
officially deprecate the old QWidgets stuff.