Deprecate AcceptHtmlDownloads property
This property is only set to true in very few
places and the downloadable items do not contain
a HTML page.

And IMO the user should always be prompted in case
there is anything unusual.

The KVHTML files used in KDE-Edu have the mimetype
of normal XML files, consequently we do not
have to worry about those.