Add support for kns:/ urls to the knewstuff-dialog tool
This adds in support for the kns URLs used in e.g. kpackagetool to
canonically reference a specific entry found on a knewstuff provider.
The result is that by passing in such a URL to the knewstuff-dialog
tool, we interpret this as wanting to show the details for that specific
entry.

The origin of this functionality was a short discussion on
https://invent.kde.org/plasma/discover/-/merge_requests/63 where the
need for such functionality became obvious.

CCBUG:430812