Refactor the KNSQuick Page header and footer for Kirigami-ness
This removes the custom header implementation used on KNSQuick's Page
component, which was originally implemented because Kirigami did not at
the time support the granularity of control needed to do the stuff we
wanted to be able to do. Now that this has changed, it was time to do a
refactor and move the things into context actions that make sense to
move there (which, essentially, is everything but the category
selection, since that one is more type dependent, and arguably a part of
the content of the page).