Put NewStuff.DialogContent's handle into the global toolbar
Without this patch, the dialog's bottom is covered by the drawer's
handle, which makes it awkward to get to specifically the close button
(but also it just looks very strange). So, using Kirigami's nice, new
canContainHandles property on the global toolbar, we put the handle up
there instead. It also looks nicer, which is... nice.