qtquickengine: Do not emit entriesChanged signal when we clear an empty list
This is called whenever we open the dialog, but we do not want to emit the signal unless
we actually clear this list. Otherwise the KCMs would reload their data and scroll back to the top.

Also we do not want the changed signal to be emitted for intermediate states,
this will only lead to noise on the KCM side and we still emit the signal when we
have the entry updated/installed.

BUG: 431568
FIXED-IN: 5.80

If we never want to scroll to the top we have to implement this on the
side of the individual KCMs.