Less risk of infinite spinner on uninstalling KPackage based things
Prior to this patch, the uninstallation function would attempt to work out if the installed item was a directory by checking for a backslash at the end of the installed files. Since we have QFileInfo, which we're already using further down in this function anyway, we might as well use the ability of that to tell us whether or not the item is a directory.

The end result is that we no longer have an infinite spinner after (successfully, as it happens) uninstalling KPackage based entries, such as Plasmoids. The symptom was most easily visible in the Plasmoid installer, which retains its engine across the entirety of a Plasma session (since the dialogue lived inside the Plasma process, and consequently never was destructed, meaning the weird state was retained).

This patch further rewords the text shown when uninstalling entries, which since it is also the text shown when installing them must be more generic. They now say some variant of "Working", to show that the entry is being worked on, which is sufficiently generic to not involve awkward formulations such as long winded "Installing or Uninstalling", or even worse "(Un)installing". At a later point in time, introducing a more full feedback system for the installation progress would be advantageous, but for a rapid fix to the immediate issue, this works.

BUG:434371