Create a NewStuff.Action component, add NewStuff.Settings global
This adds a new component called NewStuff.Action, designed to allow
for easier GHNS integration in more modern applications where specific
graphical components are not so desirable, but the direct use of e.g. a
NewStuff.Page is still too cumbersome. This action was based on a bit of
experience with implementing the folding buttons in various KCMs, and is
designed to make such work simpler in the future, and also perhaps the
work of switching to NewStuff.Page over NewStuff.Dialog (which was
always intended as a stop-gap).

This also adds a Settings utility global, which for now just holds the
information from Kiosk about whether Get Hot New Stuff is allowed, but
there are several bits of system-global usefulness that could be
exposed through this object as well.