Add functionality to have a local providers file
It is occasionally useful to be able to directly link to a feed from an
application, instead of having to pass by a remote server to resolve the
provider information, and this code allows for such a thing to be
achieved, by adding an option to knsrc which allows for adding in a
providers file next to the knsrc file.

The patch further cleans up the XmlLoader class a touch, as a result of
adding support for non-remote file locations.

Copy of the new documentation:

To use a local providers file instead of a remote one (this is useful for example in cases where you wish to use
an OPDS feed, which likely does not have a providers file, and you lack a server on which to host one of your
own), you can set <em>UseLocalProvidersFile=true</em> instead of ProvidersUrl. This will make the engine attempt
to load a file located next to your knsrc file, with the same name as that file, with .providers appended. For
example, if your configuration file is called "appname.knsrc", then your providers file should be named
"appname.providers"