Sort and capitalise items in the knewstuff-dialog tool's sidebar
Prior to this, the list of knsrc file entries in the sidebar was just
whatever sort order we got from Engine, and things without an explicit
name set in the config file were just given the filename. This makes the
list a little more more pleasant to read and navigate, by first
transforming the filename to something more akin to a real string, and
then sorting the list in a locale aware fashion.