Rework the knsrc documentation to be more central
Previously, the knsrc file documentation was duplicated (and incorrect in both prior locations),
and furthermore incomplete as it only documented part of what is possible through knsrc files,
meaning you'd have to hunt through various source files (even in some cases the implementation
and not just api docs) to find out what to do. This patch puts the documentation of this highly
important configuration file, which is the primary connection point for most users of KNewStuff,
nto the main readme file, so it is more easy to find and use.