Handle payload-only entries during update operations
During the implementation of handling the automatic payload selection system for entry updating, the situation where there were no downloaditems defined at all, or they were omitted in favour of a pre-known single payload, was left out. This patch reintroduces support for that situation (which commonly occurs in cases where the SimpleXML provider is used, for example in KStars).

This was identified as an issue by the KStars developers, and was first reported in issue #1 here (which is why there is no BUG entry below the way there usually would be for bug fixes).