Page: Remind/notify users that everything here is 3rd-party content
Otherwise it may not be obvious to some people that this stuff isn't
fully supported and tested content from KDE itself. Best to set
expectations accordingly when letting people download 3rd-party content
that modifies the appearance or behavior of the software. This is
especially relevant for things that integrate deeply into the system
like KWin effects and scripts, SDDM themes, KRunner runners, global
themes, and so on.