Engine: Do not require waiting for the providers to tell our installed entries
KNS knows which are the installed resources without needing to connect
to the server. This change addresses it so it's done in-place instead of
asking the different providers which would have told us what we already
know (note teh code on both providers is boilerplate asking the cache
too).

This allows Discover to load the providers asynchronously to the
providers startup.