Move KNS3::Action class to new KNSWidgets submodule
In KF6 the widgets should be in their own module instead of the legacy KNewStuff module
where all of the classes were before the core/gui split.

Also the 3 part of KNS3 was just a hack for co-installability added in the KDE4 times
and does not reflect any actual major version number.

Task: https://phabricator.kde.org/T15055