Unify how kpackageType is determined for the addons
Currently we have multiple sources of information, how we determine the KPackage structures:

- The X-Plasma-ServiceType property from the installed package
- The service types from the installed package
- The KPackageType from the knsrc file the app uses

With this MR we emit warnings if the package itself specifies anything different that
type from the knsrc file.
By extending the deprecation macros, the legacy code path is marked for deletion in KF6.

Task: https://phabricator.kde.org/T15136