KNSWidgets::Action add compatibility for unified entry class in KF6
This way consumers can utilize the KNSCore::Entry class name already, which
will be the unified entry class in KF6.

To avoid introducing an overloaded signal or a signal under a different name,
the signal using the new class name is only visible when building KNewStuff or building
with excluded deprecated API on the consumer side.

Relates to https://phabricator.kde.org/T15141