Engine: Default to autoconfig.kde.org for reading ProvidersUrl
With all of the issues we had recently of outdated URL values, it makes
the most sense to use the KDE autoconfig URL as a default argument.
If we ever need to change this, we only need to adjust one file and backport
one fix instead of many places from different projects and release cycles.