Add a "try again later" message for transient errors
This error dialog currently detects transient errors and changes the
icon accordingly. But it doesn't add any text. As a result, the fact
that the error is transient and the user should just try again later
remains unclear.

This commit fixes that by appending some text to indicate it as well.