Consolidate footers
Right now the page can have a footer with a combobox on it, and the
dialog that contains it can have a footer with an Upload button on it.
When both are shown at once--for example when getting new
wallpapers--the dialog shows two stacked footers each with only one
thing on it, which isn't ideal.

This comit fixes that by combining both footers into one in the page
content; now the dialog has no footer of its own.