core: Don't make URL installation a false question
There's not really an alternative, this is not supported by KNS, the
only alternative is to have the browser showing it or not then just
cancel it.

This changes the current behaviour so we include the link in the message
so it can be opened but otherwise just issues an error message and
proceeds.

This helps because:
- we don't keep the process blocked waiting for an answer
- we always error out, not only when the user decides to see it in the
browser
- we don't lie to the users making them think there's a choice