DownloadItemsSheet: Fix scrolling
OverlaySheet documentation clearly states:

>  It needs a single element declared inside, do *not* override its contentItem

…and KNewStuff DownloadItemsSheet.qml happily does just that:

> Kirigami.OverlaySheet {
>     id: component
>
>     //...
>
>     contentItem: QtLayouts.ColumnLayout {
>         QtControls.ScrollView {
>             // ...

Second part of the problem is that the docs suggests using either a
ColumnLayout or a ListView as that single element. But not combining
them one inside another, and especially not a ListView wrapped in an
extra ScrollView inside a ColumnLayout. Moving extra top items into a
ListView's inline header, on the other hand, works great, and fully
fixes this scrolling issue.

Note: In order to backport to kf5 branch, replace `icon.name:` in
Kirigami.BasicListItem delegate with older `icon:` counterpart.

BUG: 448800
FIXED-IN: 5.104