Install headers into fully module-namespaced subdirs
The basic layout of include dirs of libs from KF modules is:
<visibility-layer>/<C++ namespace>/<class headers>
where visibility-layer is KF6/LibraryBaseName and the dir which
is listed as the include dir to use for a library.
This ensures that only headers will be seen when using the expected
include statements (like #include <CppNameSpace/Class>) which belong
to the library one wants to link to.