Installation: Read KPackageStructure key from knsrc file
For packages and the package structures, this key is also called KPackageStructure.
Consequently, having the same key for semantically the same value reduces the risk of errors
and makes things more consistent.

See https://invent.kde.org/teams/frameworks-devs/kf6-workboard/-/issues/30#note_629405

When we have the first ever KF6 release, the old key should be removed.
But during the transition period, it is useful to read both.