KNSWidgets::Button: Expose setConfigFile back as a proper method
With qdesigner, the property is translated by default.
Changing this is not possible, unless we modify the C++ plugin
generated by ECMAddQtDesignerPlugin.

Considering this, having consumers manually call this method is an
easier solution than fixing the qdesigner edgecase.

This is also what I did in the Konsersation port back in the day.