Simplify Question related code
For internal API, we do not need the helper and can store the question as a simple member variable
Providing a default implementation for the QuestionManager::askQuestion
method doesn't make any sense. Thus we should enforce consumers to
overide this key.

Instead of an empty d-ptr implementation, a void pointer is kept for future BIC.