QML: Simplify Error Message and fix its dimensions
We have Kirigami.PromptDialog now, so no need to generate a custom
MessageBoxSheet. This is a more proper control for showing errors
and fixes the small dimensions of the previous versions.

While at it also use Kirigami.SelectableLabel, remove import versions,
and fix the link signal handler by setting and explicit argument.

BUG: 476150
FIXED-IN: 6