QuickEngine: Fix Installed and Updateable filters
When selecting the "Show only updateable entries" filter, we
must invalidate the cache the same way as we do for the "installed"
filter, since this is a subset of that one.

Then, the attica provider launches new jobs to check if the items
are updateable, which produces a lot of `DetailsLoadedEvent`, one
for each checked entry.

Filter out those events when they no longer apply to the new state
of the filter.

BUG: 474548
FIXED-IN: 6.0