delegates: Fix thumbnail placeholder icon position
The thumbnail placeholder, shown when the item has no previews,
assumed the thumbnail would cover the whole delegate, so when it's
not the case, it can show in unwanted places.

To fix this, add a `thumbnailArea` property which allows the
derived delegates to set it where the actual thumbnail is placed.

The previous item named `thumbnailArea` is renamed to `contentArea`
to better reflect what it does and avoid confusion.