[qtquick/action] Add API to set transient parent for dialog
We want the dialog to have a proper parent window set. Usually
Qt automatically obtains one from the parent item. This doesn't
work here however since the dialog's loader is not part of the item
hierarchy. This causes the window to not be shown since Qt 6.7.

See QTBUG-121197

Merely adding the property and setting it to null makes sure the
window can be shown. Users of the API should however pass a parent window
so that the dialog has a parent set.

BUG: 479816