Explicitly set QQC2.Overlay.overlay as a parent for popups
Because that's how popups work in QtQuick.Controls. And that's what goes
wrong in Kirigami with its "implicit" bindings directly inside Dialog
and OverlaySheet types. Because QQC2.Overlay.overlay has to be queried
on a real Item which already belongs to some window; if you query it on
a Popup itself you get's get null overlay, but the Popup are usually
declared as child objects of some other Item which which case said Item
would become not only QObject::parent but also a visual Item parent of
a Popup, thus accidentally making `QQC2.Overlay.overlay` work -- only
to immediately reassign parent Item to an overlay itself. This sounds
like a rather unintentional loophole rather than a well-thought
architectural design, so let's avoid that pattern, and just bind a
parent normally.