support new version field on DownloadLinkInformation
this shuffles the internals of Entry such that we get away without
having to introduce new public API that would ultimately be of no use
since the version should (currently) only matter inside the installation
use case and that is encapsulated inside the Transaction.

in particular the public struct DownloadLinkInformation is of concern
since we'd need to create a new public V2 variant and then juggle V1 and
V2 at the same time in the API. lots of API noise for no gain.

so, EntryPrivate and DownloadLinkInformationV2Private are now in an own
header and directly utilized by Transaction rather than going through
public API. internally Entry only holds DownloadLinkInformationV2Private
instances. to keep compatibility with V1 API we synthesize V1 objects
from V2 objects on-demand and vice versa (i.e. in the public getter and
setter)