Use Tiles view by default
Currently the default was unclear: Page.qml defaulted to Tiles, but the
code elsewhere on the page marked the default as something else in its
switch statement (maybe just an error?) and also the the default value
was overridden in Action.qml and Button.qml.

This situation was incoherent, so let's instead settle on Tiles for
everything. It's more compact and lets the user see more content at a
time.