Remove user-facing view switcher
It's reasonable for the implementation to choose a view mode; for
example some benefit from showing large images, and some don't. But it's
super overkill to offer this possibility to the user.

Remove the user-facing view switcher and make it only something the
implementation gets to choose.

This has the side benefit of removing content from the toolbar which
makes the search field less likely to get stuffed into an overflow menu
with small window sizes or languages with long text.