Remove Icons view
Large Preview view is useful for things like wallpapers where the image
is the most important thing to see.

Tiles view is useful for everything else.

Icons view is useful for nothing, so let's delete its implementation.
Since this is a framework, we can't remove the API to request it, so
instead any such requests will return Tiles instead.