the grand API refactor of 2024
searchrequest wanted to grow an id but it couldn't because it was a
public struct and compatibility rigging also was difficult because
Provider is a virtual non-leaf class. apol and I decided it'd be better
to just fix what's wrong here.

the primary change here is that we are retracting (or rather
deprecating, for now) a whole lot of public interfaces that aren't
actually needed by the outside just now. the goal here is a leaner ABI
and better API (WRT encapsulation of internals).

Provider has been entirely deprecated. This was a backwards
compatibility trap, leaking abstraction (the callers got direct access
to otherwise internal provider objects), didn't need to be public, and
technically consituted support for third party Provider instances.

It has been replaced by ProviderCore (the public frontend API) and
ProviderBase (the private backend API). The public frontend API had
pretty much all its interfaces removed as they appear to be unused. The
backend API meanwhile has been adopted by all internal providers and is
more or less simply the old Provider - with it now being private we no
longer have to worry about over-exposing API since we can just remove
stuff when it becomes unnecessary.

All public structs have been turned into classes with d pointers. They
have moved scope to differentiate them from the old structs. The structs
have all been deprecated.

As part of this the API has undergone some tightening where API seemed
unused by the outside. It is instead being used internally via d pointer
friendships. That is to say that a quick class that needs access to e.g.
searchpresets from the EngineBase now obtains it via EngineBasePrivate
instead of offering public API. A tigther public API means less stuff to
worry about when making changes.

This entire change set is fully backwards compatible on an ABI level.
And even improves reliability because the new internal id of the
SearchRequests allow better handling of multiple inflight requests that
look the same.
On the runtime side this is not necessarily the case. Technically one
could have previously implemented a Provider and supplied it to the
EngineBase. This was an unintended feature and the code path is now dead
at runtime. While this is a feature we could support it would need
greater care with regards to interface separation and API exposure.

As per usual it is greatly advised to port away from deprecated API.
Doing so mostly involves changing from Provider to ProviderCore (in so
far as one even needs to use the Provider directly -- protip: you
probably don't), changing the scopes of SearchRequest et al, and
changing struct member access to function calls.