cache: become a facade for Cache2
I hadn't realized this originally but Cache isn't actually a cache but a
persistent registry. As such it makes no sense to have two objects that
do the same thing. Instead redirect the internals of Cache to the
relevant instance of Cache2

the way this is achieved is through the actual cache aspect of the
registry. the Cache ctor calls Cache2::getCache to get the unique
instance of the appName's cache for delegating to. the ABI remains
entirely unchanged

enginebase now treats legacyCache as a facade that doesn't need calling
into