Be clear about where content comes from
Right now there's no communication about the fact that this dialog
reaches out to a 3rd-party service, leading people to conclude that KDE
owns and controls it, when neither are true.

Let's address that by adding some branding for the service itself
(Pling) to let users know they're reaching outside of KDE when they use
this.